backend/requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio>=3.7.1,<4.0.0
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.1.0.post1
fastapi==0.109.2
fastapi-cli==0.0.7
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.4.8
pycparser==2.22
pydantic==2.6.1
pydantic-core==2.14.5
pydantic[email]
python-jose[cryptography]==3.3.0
Pygments==2.19.1
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.1
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.27
starlette>=0.27.0,<0.28.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.27.1
watchfiles==1.0.5
websockets==15.0.1
openai-whisper==20240930
torch==2.2.0
ffmpeg-python==0.2.0
aiofiles==23.2.1
python-magic==0.4.27
numpy==2.2.0
transformers==4.37.2
psutil==5.9.8
spacy==3.7.4
symspellpy==6.7.7
hf_xet==1.1.3
huggingface_hub[hf_xet]==0.32.4
fr-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.7.0/fr_core_news_sm-3.7.0-py3-none-any.whl
